feat: add configuration
