+++ nodemon.json
@@ -1,0 +1,8 @@
+{
+  "watch": ["src"],
+  "ext": "ts",
+  "ignore": ["*.test.ts"],
+  "execMap": {
+    "ts": "ts-node"
+  }
+}
